requirements.txt
# FastAPI and related libraries
fastapi # fastapi
starlette # starlette
uvicorn
pydantic # pydantic

# Environment variables
python-dotenv # dotenv

# JWT and cryptography
python-jose[cryptography] # jose

# Testing
pytest # pytest
pytest-asyncio
protobuf

# HTTP client
httpx
attrs
python-dateutil

# Database
asyncmy
sqlmodel # sqlmodel
SQLAlchemy # sqlalchemy
